requirements.txt
scrapetube
yt-dlp
python-pptx
python-dotenv
matplotlib
requests
numpy
